Rename example to simple-native
