+++ docs/src/examples/index.md
@@ -3,3 +3,6 @@
 - [Hello World](./hello-indexer.md)
   - A simple application representing the "Hello world" version of a Fuel indexer
   - This example includes WASM execution
+- [Simple Native](./simple-native.md)
+  - A simple example of an indexer within a full Fuel project. 
+  - Also demonstrates native execution.
