+++ examples/simple-native/programs/counter-rs/Cargo.toml
