+++ examples/simple-native/programs/counter/src/main.sw
