+++ examples/simple-native/schema/counter.graphql
